email-processor/requirements.txt
pydantic-settings>=2.0.0
pyyaml>=6.0
boto3>=1.34.0
jira>=3.8.0
msal>=1.28.0
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.9
weasyprint>=61.0
